fix: resolve build failure and improve workflow formatting
🔧 **Build Fixes:**
- Fix JSX structure error in WizardShell.tsx (lines ~910-930)
- Add missing vi import in test setup file
- Resolve deployment failure from malformed Average Net Fee field

📝 **Workflow Improvements:**
- Improve template string formatting in GitHub Actions workflows
- Enhance readability of automated comment templates
- Standardize formatting across wiki-automation, project-automation,
  ai-code-review, and auto-issue-creation workflows

🧩 **New Components:**
- Add modular Wizard components for better code organization
- Implement ExistingStoreSection and NewStoreSection
- Add StrategicAnalysis component with performance insights
- Create shared calculations and types for wizard flow

✅ **Impact:**
- Resolves preview deployment failures
- Improves maintainability of GitHub workflows
- Better code organization for wizard functionality
- Fixes testing infrastructure issues

Closes build error from commit 20ba82c on feat/expenses-dual-entry branch.
