+++ src/components/Wizard/NewStoreSection.tsx
@@ -1,0 +1,221 @@
+// NewStoreSection.tsx - Target Performance Goals for new stores
+// Manual entry forecasting without growth slider
+
+import React from 'react'
+import type { WizardSectionProps } from './types'
+import { calculateNetIncome, formatCurrency, parseCurrencyInput } from './calculations'
+
+export default function NewStoreSection({ answers, updateAnswers, region }: WizardSectionProps) {
+  return (
+    <>
+      {/* Information Banner */}
+      <div style={{ 
+        padding: '1rem', 
+        backgroundColor: '#f0f9ff', 
+        border: '1px solid #0ea5e9', 
+        borderRadius: '6px',
+        marginBottom: '1rem'
+      }}>
+        <div style={{ fontWeight: 'bold', color: '#0369a1', marginBottom: '0.5rem' }}>
+          🏪 New Store Setup - Forecasting
+        </div>
+        <div className="small" style={{ color: '#0369a1' }}>
+          Set your target performance goals below. These will be used for business planning and can be adjusted as you learn more about your market.
+        </div>
+      </div>
+
+      {/* Target Performance Goals Box */}
+      <div style={{
+        marginBottom: '1rem',
+        padding: '1.5rem',
+        backgroundColor: '#f8fafc',
+        border: '1px solid #e2e8f0',
+        borderRadius: '8px'
+      }}>
+        <div>
+          <h4 style={{ 
+            margin: '0 0 0.75rem 0', 
+            fontSize: '1rem', 
+            fontWeight: 600, 
+            color: '#374151',
+            borderBottom: '2px solid #059669',
+            paddingBottom: '0.25rem'
+          }}>
+            Target Performance Goals
+          </h4>
+
+          {/* Average Net Fee - Manual entry */}
+          <div style={{ marginBottom: '0.75rem' }}>
+            <div style={{ display: 'flex', alignItems: 'center', gap: '0.75rem', marginBottom: '0.25rem' }}>
+              <label style={{ minWidth: '120px', fontWeight: 500 }}>Average Net Fee</label>
+              <div style={{ display: 'flex', alignItems: 'center', gap: '0.25rem' }}>
+                <span style={{ fontWeight: 500, color: '#6b7280' }}>$</span>
+                <input
+                  type="text"
+                  placeholder="e.g., 130"
+                  value={formatCurrency(answers.avgNetFee)}
+                  onChange={e => {
+                    const newValue = parseCurrencyInput(e.target.value)
+                    updateAnswers({ avgNetFee: newValue })
+                  }}
+                  style={{ 
+                    width: '140px', 
+                    textAlign: 'right', 
+                    border: '1px solid #d1d5db', 
+                    borderRadius: '4px', 
+                    padding: '0.5rem' 
+                  }}
+                />
+              </div>
+            </div>
+            <div className="small" style={{ marginTop: '0.25rem', marginLeft: '100px', opacity: 0.7 }}>
+              Your target average net fee per return
+            </div>
+          </div>
+
+          {/* Tax Prep Returns - Manual entry */}
+          <div style={{ marginBottom: '0.75rem' }}>
+            <div style={{ display: 'flex', alignItems: 'center', gap: '0.75rem', marginBottom: '0.25rem' }}>
+              <label style={{ minWidth: '120px', fontWeight: 500 }}>Tax Prep Returns</label>
+              <div style={{ display: 'flex', alignItems: 'center', gap: '0.25rem' }}>
+                <span style={{ fontWeight: 500, color: '#6b7280' }}>#</span>
+                <input
+                  type="number"
+                  placeholder="e.g., 1,680"
+                  value={answers.taxPrepReturns || ''}
+                  onChange={e => updateAnswers({ 
+                    taxPrepReturns: parseFloat(e.target.value) || undefined 
+                  })}
+                  style={{ 
+                    width: '140px', 
+                    textAlign: 'right', 
+                    border: '1px solid #d1d5db', 
+                    borderRadius: '4px', 
+                    padding: '0.5rem' 
+                  }}
+                />
+              </div>
+            </div>
+            <div className="small" style={{ marginTop: '0.25rem', marginLeft: '100px', opacity: 0.7 }}>
+              Your target number of tax returns
+            </div>
+          </div>
+
+          {/* TaxRush Returns (Canada only) */}
+          {region === 'CA' && (
+            <div style={{ marginBottom: '0.75rem' }}>
+              <div style={{ display: 'flex', alignItems: 'center', gap: '0.75rem', marginBottom: '0.25rem' }}>
+                <label style={{ minWidth: '120px', fontWeight: 500 }}>TaxRush Returns</label>
+                <div style={{ display: 'flex', alignItems: 'center', gap: '0.25rem' }}>
+                  <span style={{ fontWeight: 500, color: '#6b7280' }}>#</span>
+                  <input
+                    type="number"
+                    placeholder="0"
+                    value={answers.taxRushReturns || ''}
+                    onChange={e => updateAnswers({ 
+                      taxRushReturns: parseFloat(e.target.value) || undefined 
+                    })}
+                    style={{ 
+                      width: '140px', 
+                      textAlign: 'right', 
+                      border: '1px solid #d1d5db', 
+                      borderRadius: '4px', 
+                      padding: '0.5rem' 
+                    }}
+                  />
+                </div>
+              </div>
+              <div className="small" style={{ marginTop: '0.25rem', marginLeft: '100px', opacity: 0.7 }}>
+                Your target TaxRush returns (Canada only)
+              </div>
+            </div>
+          )}
+
+          {/* Gross Tax Prep Fees - Auto-calculated */}
+          <div style={{ marginBottom: '0.75rem' }}>
+            <div style={{ display: 'flex', alignItems: 'center', gap: '0.75rem', marginBottom: '0.25rem' }}>
+              <label style={{ minWidth: '120px', fontWeight: 500 }}>Gross Tax Prep Fees</label>
+              <div style={{ display: 'flex', alignItems: 'center', gap: '0.25rem' }}>
+                <span style={{ fontWeight: 500, color: '#6b7280' }}>$</span>
+                <input
+                  type="text"
+                  value={(() => {
+                    if (answers.avgNetFee && answers.taxPrepReturns) {
+                      const grossFees = Math.round(answers.avgNetFee * answers.taxPrepReturns)
+                      return grossFees.toLocaleString()
+                    }
+                    return ''
+                  })()}
+                  readOnly
+                  style={{ 
+                    width: '140px', 
+                    textAlign: 'right', 
+                    border: '1px solid #d1d5db', 
+                    borderRadius: '4px', 
+                    padding: '0.5rem', 
+                    backgroundColor: '#f9fafb' 
+                  }}
+                />
+              </div>
+            </div>
+            <div className="small" style={{ marginTop: '0.25rem', marginLeft: '100px', opacity: 0.7 }}>
+              Auto-calculated: Average Net Fee × Tax Prep Returns
+            </div>
+          </div>
+
+          {/* Total Expenses - Auto-calculated with override */}
+          <div style={{ marginBottom: '0.75rem' }}>
+            <div style={{ display: 'flex', alignItems: 'center', gap: '0.75rem', marginBottom: '0.25rem' }}>
+              <label style={{ minWidth: '120px', fontWeight: 500 }}>Total Expenses</label>
+              <div style={{ display: 'flex', alignItems: 'center', gap: '0.25rem' }}>
+                <span style={{ fontWeight: 500, color: '#6b7280' }}>$</span>
+                <input
+                  type="text"
+                  value={(() => {
+                    if (answers.avgNetFee && answers.taxPrepReturns) {
+                      const grossFees = answers.avgNetFee * answers.taxPrepReturns
+                      const expenses = Math.round(grossFees * 0.76)
+                      return expenses.toLocaleString()
+                    }
+                    return ''
+                  })()}
+                  onChange={e => {
+                    const newValue = parseCurrencyInput(e.target.value)
+                    updateAnswers({ projectedExpenses: newValue })
+                  }}
+                  style={{ 
+                    width: '140px', 
+                    textAlign: 'right', 
+                    border: '1px solid #d1d5db', 
+                    borderRadius: '4px', 
+                    padding: '0.5rem' 
+                  }}
+                />
+              </div>
+            </div>
+            <div className="small" style={{ marginTop: '0.25rem', marginLeft: '100px', opacity: 0.7 }}>
+              Industry standard: 76% of Gross Tax Prep Fees (you can override)
+            </div>
+          </div>
+
+          {/* Target Net Income Summary */}
+          {answers.avgNetFee && answers.taxPrepReturns && (
+            <div style={{ 
+              padding: '0.5rem', 
+              backgroundColor: '#e0f2fe', 
+              borderRadius: '4px',
+              fontWeight: 600,
+              fontSize: '0.9rem',
+              color: '#0369a1'
+            }}>
+              Target Net Income: ${calculateNetIncome(answers).toLocaleString()}
+              <div style={{ fontSize: '0.8rem', fontWeight: 'normal' }}>
+                Net Margin: 24% (industry standard)
+              </div>
+            </div>
+          )}
+        </div>
+      </div>
+    </>
+  )
+}
