+++ src/test/setup.ts
@@ -1,4 +1,5 @@
 import '@testing-library/jest-dom'
+import { vi } from 'vitest'
 
 // Mock localStorage for tests
 const localStorageMock = (() => {
